[KZL-375] Server Controller
