+++ src/Kuzzle.php
@@ -2,7 +2,6 @@
 
 namespace Kuzzle;
 
-use DateTime;
 use ErrorException;
 use Exception;
 use InvalidArgumentException;
@@ -14,6 +13,7 @@ use Kuzzle\Util\CurlRequest;
 use Kuzzle\Security\Security;
 use Kuzzle\Security\User;
 use Kuzzle\Bulk;
+use Kuzzle\Server;
 
 /**
  * Class Kuzzle
@@ -89,6 +89,11 @@ class Kuzzle
      */
     public $bulk;
 
+    /**
+     * @var Server Kuzzle's Server controller
+     */
+    public $server;
+
 
     /**
      * Kuzzle constructor.
@@ -127,6 +132,7 @@ class Kuzzle
 
         // API Controllers
         $this->bulk = new Bulk($this);
+        $this->server = new Server($this);
 
         return $this;
     }
@@ -253,24 +259,6 @@ class Kuzzle
     }
 
     /**
-     * Kuzzle monitors active connections, and ongoing/completed/failed requests.
-     * This method returns all available statistics from Kuzzle.
-     *
-     * @param array $options Optional parameters
-     * @return array[] each one of them being a statistic frame
-     */
-    public function getAllStatistics(array $options = [])
-    {
-        $response = $this->query(
-            $this->buildQueryArgs('server', 'getAllStats'),
-            [],
-            $options
-        );
-
-        return $response['result']['hits'];
-    }
-
-    /**
      * @return string
      */
     public function getDefaultIndex()
@@ -319,54 +307,6 @@ class Kuzzle
         );
 
         return $response['result']['hits'];
-    }
-
-    /**
-     * Retrieves information about Kuzzle, its plugins and active services.
-     *
-     * @param array $options Optional parameters
-     * @return array containing server information
-     */
-    public function getServerInfo(array $options = [])
-    {
-        $response = $this->query(
-            $this->buildQueryArgs('server', 'info'),
-            [],
-            $options
-        );
-
-        return $response['result']['serverInfo'];
-    }
-
-    /**
-     * Kuzzle monitors active connections, and ongoing/completed/failed requests.
-     * This method allows getting either the last statistics frame,
-     * or a set of frames starting from a provided timestamp.
-     *
-     * @param string $timestamp Optional starting time from which the frames are to be retrieved
-     * @param array $options Optional parameters
-     * @return array[] containing one or more statistics frame(s)
-     */
-    public function getStatistics($timestamp = '', array $options = [])
-    {
-        $data = [];
-
-        if (empty($timestamp)) {
-            $action = 'getLastStats';
-        } else {
-            $action = 'getStats';
-            $data['body'] = [
-                'startTime' => $timestamp
-            ];
-        }
-
-        $response = $this->query(
-            $this->buildQueryArgs('server', $action),
-            $data,
-            $options
-        );
-
-        return empty($timestamp) ? [$response['result']] : $response['result']['hits'];
     }
 
     /**
@@ -502,23 +442,6 @@ class Kuzzle
         }
 
         return $memoryStorage;
-    }
-
-    /**
-     * Retrieves the current Kuzzle time.
-     *
-     * @param array $options Optional parameters
-     * @return DateTime
-     */
-    public function now(array $options = [])
-    {
-        $response = $this->query(
-            $this->buildQueryArgs('server', 'now'),
-            [],
-            $options
-        );
-
-        return new DateTime('@' . round($response['result']['now'] / 1000));
     }
 
     /**
