+++ tests/KuzzleTest.php
@@ -304,54 +304,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($createIndexResponse, $response);
     }
 
-    public function testGetAllStatistics()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_getAllStats',
-            'request' => [
-                'action' => 'getAllStats',
-                'controller' => 'server',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $getAllStatisticsResponse = ['hits' => []];
-        $httpResponse = [
-            'error' => null,
-            'result' => $getAllStatisticsResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->getAllStatistics($options);
-
-        $this->assertEquals($getAllStatisticsResponse['hits'], $response);
-    }
-
     public function testGetMyRights()
     {
         $url = self::FAKE_KUZZLE_HOST;
@@ -398,151 +350,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
         $response = $kuzzle->getMyRights($options);
 
         $this->assertEquals($getMyRightsResponse['hits'], $response);
-    }
-
-    public function testGetServerInfo()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_serverInfo',
-            'request' => [
-                'action' => 'info',
-                'controller' => 'server',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $getServerInfoResponse = ['serverInfo' => []];
-        $httpResponse = [
-            'error' => null,
-            'result' => $getServerInfoResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        $response = $kuzzle->getServerInfo($options);
-
-        $this->assertEquals($getServerInfoResponse['serverInfo'], $response);
-    }
-
-    public function testGetLastStatistics()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_getLastStats',
-            'request' => [
-                'action' => 'getLastStats',
-                'controller' => 'server',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $getLastStatsResponse = [];
-        $httpResponse = [
-            'error' => null,
-            'result' => $getLastStatsResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->getStatistics('', $options);
-
-        $this->assertEquals([$getLastStatsResponse], $response);
-    }
-
-    public function testGetStatistics()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-        $statsTime = time();
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_getStats',
-            'request' => [
-                'action' => 'getStats',
-                'controller' => 'server',
-                'volatile' => [],
-                'body' => [
-                    'startTime' => $statsTime
-                ],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'POST',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $getStatsResponse = ['hits' => []];
-        $httpResponse = [
-            'error' => null,
-            'result' => $getStatsResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->getStatistics($statsTime, $options);
-
-        $this->assertEquals($getStatsResponse['hits'], $response);
     }
 
     public function testListAllCollections()
@@ -902,55 +709,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals($logoutResponse, $response);
         $this->assertAttributeEquals(null, 'jwtToken', $kuzzle);
-    }
-
-    public function testNow()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_now',
-            'request' => [
-                'action' => 'now',
-                'controller' => 'server',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        // mock response (timestamp is given in milliseconds)
-        $logoutResponse = ['now' => (time()) * 1000];
-        $httpResponse = [
-            'error' => null,
-            'result' => $logoutResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->now($options);
-
-        $this->assertInstanceOf('DateTime', $response);
-        $this->assertEquals($logoutResponse['now'] / 1000, $response->getTimestamp());
     }
 
     public function testRefreshIndex()
